Protect map geometry and renderer teardown
Use long-safe map bounds, validate patch IDs, and respect the protocol bundle cap. Serialize clear with restart, restore removed-viewer state, and handle main-thread shutdown. Document that synchronous clear and viewer snapshots bypass the delta byte budget.

Validation: final combined tree passed spotlessApply, build, coverageLint, Javadoc and E2E source compilation on Linux/GraalVM 25. See sept24-finish-build3.log and sept24-finish-validation in the pass-2 evidence directory. Intermediate commits were not built independently; the logical series shares test fixtures and API changes.

+++ mcav-bukkit/src/main/java/me/brandonli/mcav/bukkit/media/map/MapTilePatch.java
@@ -56,7 +56,7 @@ public final class MapTilePatch {
 
     final boolean negativeOrigin = x < 0 || y < 0;
     final boolean emptySize = width <= 0 || height <= 0;
-    final boolean exceedsMap = x + width > MapLayout.MAP_SIZE || y + height > MapLayout.MAP_SIZE;
+    final boolean exceedsMap = (long) x + width > MapLayout.MAP_SIZE || (long) y + height > MapLayout.MAP_SIZE;
     final boolean insideMap = !negativeOrigin && !emptySize && !exceedsMap;
     Preconditions.checkArgument(insideMap, "Patch is outside of the map bounds");
